backend/ai_startup_jobs/requirements.txt
# Core
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
python-dotenv>=1.0.0

# Playwright (headless browser for dynamic sites)
playwright>=1.40.0
playwright-stealth>=1.0.6

# Apify (LinkedIn scraping)
apify-client>=1.7.0

# Google Sheets
gspread>=6.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0

# Scheduling
schedule>=1.2.0

# Utilities
lxml>=4.9.0
fake-useragent>=1.4.0
tenacity>=8.2.0
